LayoutTests/http/tests/w3c/html/semantics/embedded-content/the-img-element/srcset/select-an-image-source-expected.txt

            

FAIL "data:,a 1x, data:,b 1x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a , data:,b 1x" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1x, data:,b" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1w, data:,b 1w" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1w 1h, data:,b 1w" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1w, data:,b 1w 1h" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1w 1h, data:,b 1w 2h" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1w 2h, data:,b 1w 1h" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a , data:,b" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1w, data:,b 1x" sizes="1px" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1x, data:,b 1w" sizes="1px" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 1w, data:,b 2x" sizes="0.5px" assert_equals: expected (string) "data:,a" but got (undefined) undefined
FAIL "data:,a 2x, data:,b 1w" sizes="0.5px" assert_equals: expected (string) "data:,a" but got (undefined) undefined
